Optimized and shrinked output files size. Also have made some changes

+++ files2py.py
@@ -1,6 +1,7 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 
+import codecs
 from os.path import isfile as file_exists
 
 
@@ -14,7 +15,10 @@ while True:
         break
     else:
         if file_exists(path):
-            paths.append(path)
+            if path not in paths:
+                paths.append(path)
+            else:
+                print('File:', path, 'has already been given!')
         else:
             print('File:', path, 'is not found!')
 
@@ -22,7 +26,13 @@ if len(paths) != 0:
     fname = input('Input filename: ')
     fname = 'None.py' if fname == '' else fname
     fname += '.py'if fname[-3:] != '.py' else ''
+    if file_exists(fname):
+        if input("File: " + fname + " already exists. Do you want to rewrite it (y/n)?").lower() != 'y':
+            exit()
     with open(fname, 'wb') as script_file:
+        script_file.write("from os.path import isfile as file_exists\n".encode())
+        script_file.write("from codecs import decode as codecs_decode\n".encode())
+        script_file.write("from binascii import unhexlify as binascii_unhexlify\n\n".encode())
         script_file.write("file_names = [".encode())
         for path in paths:
             while path.find('/') != -1:
@@ -31,17 +41,20 @@ if len(paths) != 0:
         script_file.write("]\n\n".encode())
         script_file.write("byte_files = [".encode())
         for file in paths:
-            bytes_of_file = []
-            with open(file, 'rb') as file_in_bytes:
-                for byte in file_in_bytes:
-                    bytes_of_file.append(byte)
-            script_file.write((str(bytes_of_file) + ", ").encode())
+            with open(file, 'rb') as reading_file:
+                bytes_of_file = reading_file.read()
+            bytes_of_file = codecs.encode(bytes_of_file, 'zip').hex()
+            script_file.write(('"' + str(bytes_of_file) + '",\n').encode())
 
         script_file.write("]\n\n".encode("utf-8"))
         script_file.write("for i in range(len(file_names)):\n".encode())
+        script_file.write("    if file_exists(file_names[i]):\n".encode())
+        script_file.write("        if input('File: ' + file_names[i] + ' already exists. Do you want to rewrite it (y/n)?').lower() != 'y':\n".encode())
+        script_file.write("            print('File: ' + file_names[i] + ' has been skipped!')\n".encode())
+        script_file.write("            continue\n".encode())
         script_file.write("    with open(file_names[i], 'wb') as new_file:\n".encode())
-        script_file.write("        for byte_file in byte_files[i]:\n".encode())
-        script_file.write("            new_file.write(byte_file)\n".encode())
+        script_file.write("        bytes_of_file = codecs_decode(binascii_unhexlify(byte_files[i]), 'zip')\n".encode())
+        script_file.write("        new_file.write(bytes_of_file )\n".encode())
 else:
     print('No path was given!')
 input('Press Enter to finish script...')
